Revert "Update backend structure"
This reverts commit 74539ab0341f66155144781f268dee4ddb44f27c.
