--- backend/run.py
@@ -1,4 +1,0 @@
-import uvicorn
-
-if __name__ == "__main__":
-    uvicorn.run("main:app", port=8000, reload=True)
